Added the snap = "out" argument to extend
Maybe not necessary, but this ensures the extend function *always* extends
the grid. No difference in this example, but in some cases it could result
in the grid remaining the same size in one or more dimensions, if a point
falls exactly on the edge of the grid. That's suprising and easy to miss,
so I prefer to do this explicitly.